Added window limitations / startup functionality
In this commit I have changed the way that R6DB opens so that the mainWindow has to wait 2000ms until the "load box" window is destroyed to open, which should give enough time for the local db to supply the user with the appropriate information needed. Also, as a personal rule I have added window limitations to the mainWindow, These limitations are based on past experiences when having the app at 7680px x 4320px causes system stability issues. I have also updated the package.json file by adding a "repository" field which can eventually be used by npm packages such as electron-updater to fetch new app versions from a readily available github releases server. Another update to the package.json is  the "name" field, I have switched it from "app2" to "r6db" which when changed is mirrored to package-lock.json